refactor(frontend/cart): apply FE refactoring rules to cart module
- Separate CartTable and CartSidebar sub-components into their own directories with dedicated CSS
- Hoist useCart state to page.tsx to share checkout/amount status
- Update architecture rules to include Hook State Lifting (Rule 6)
