+++ docs/Architecture/Order_FE_Refactoring_Rules.md
@@ -27,3 +27,8 @@
 - **규칙**: 단일 컴포넌트 기능 구현을 위해 관련된 `.tsx` 파일이 2개 이상 생성되는 경우, 내부 하위 폴더를 생성하여 그룹화합니다.
 - **적용 패턴**: 단일 `_components/ComponentA.tsx` 파일 내에서 관리하기 벅차 분할이 필요한 경우, `_components/ComponentA/` 폴더를 생성하고 그 안에 관련된 `.tsx` 파일들과 전용 `.module.css`를 함께 응집시킵니다.
 - **이유**: 폴더 내부 구조가 지저분해지는 것을 막고 하나의 컴포넌트 단위를 통째로 관리 가능한 패키지/모듈처럼 작동하게 합니다.
+
+## 6. 상태 공유 훅의 위치 및 상태 끌어올리기 (State Lifting)
+- **규칙**: 여러 하위 컴포넌트가 동일한 상태(State)를 공유하여 동작해야 하는 훅(Hook)의 경우, 반드시 이들을 모두 포함하는 **공통 부모(일반적으로 `page.tsx`)와 동일한 디렉터리 레벨**에 위치해야 합니다.
+- **적용 패턴**: `CartTable`과 `CartSidebar`가 동일한 결제 금액 및 장바구니 리스트(`useCart`)를 바라봐야 할 때, `useCart.ts`를 개별 컴포넌트 폴더 하위로 내리지 않고 루트 디렉터리에 둔 뒤 부모 래퍼에서 호출하여 `Props`로 하향 전달(Drilling)합니다.
+- **이유**: 커스텀 훅 내부에 선언된 로컬 상태(`useState` 등)가 각각의 자식 컴포넌트에서 독립적으로 호출될 경우 상태 파편화(데이터 비동기화) 에러가 발생하기 때문입니다. 상태 공유가 필요한 훅은 **실제 상태가 통합 관리되어야 하는 가장 높은 조상 위치**에 종속되어야 합니다.
