+++ frontend/src/app/cart/_components/CartTable/CartTable.module.css
@@ -1,37 +1,4 @@
-/* Cart Page Styles - Using global CSS variables */
-
-.container {
-  display: flex;
-  gap: var(--space-32);
-  align-items: flex-start;
-  padding: var(--space-100) var(--space-160);
-  width: 100%;
-  min-height: 100vh;
-  box-sizing: border-box;
-}
-
-.content {
-  display: flex;
-  gap: var(--space-32);
-  width: 100%;
-  flex: 1;
-}
-
-.main {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-32);
-  flex: 1;
-  min-width: 0;
-}
-
-.title {
-  font: var(--text-h1-bold);
-  color: var(--color-text-gray-900);
-  margin: 0;
-  height: 36px;
-}
-
+/* CartTable Styles */
 /* Cart Table Styles */
 .table {
   display: flex;
@@ -224,141 +191,4 @@
   transition: color var(--transition), transform var(--transition), opacity var(--transition);
 }
 
-/* Order Sidebar Styles */
-.sidebar {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-32);
-  width: 240px;
-  flex-shrink: 0;
-}
-
-.sidebarTitle {
-  font: var(--text-h1-bold);
-  color: var(--color-text-gray-900);
-  margin: 0;
-  height: 36px;
-}
-
-.summary {
-  background-color: var(--color-surface);
-  border-radius: var(--radius-lg);
-  padding: var(--space-24);
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-16);
-}
-
-.summaryDetails {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-12);
-  height: 120px;
-}
-
-.summaryRow {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.summaryRow :global(.label) {
-  font: var(--text-body-medium);
-  color: var(--color-text-gray-500);
-}
-
-.summaryRow :global(.value) {
-  font: var(--text-body-medium);
-  color: var(--color-text-gray-900);
-}
-
-.summaryRow :global(.discount) {
-  color: var(--color-error);
-}
-
-.divider {
-  height: 1px;
-  background-color: var(--color-text-gray-500);
-  width: 100%;
-}
-
-.total {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.total :global(.label) {
-  font: var(--text-body-medium);
-  color: var(--color-text-gray-900);
-}
-
-.total :global(.total) {
-  font: var(--text-body-medium);
-  color: var(--color-text-gray-900);
-  font-weight: 700;
-}
-
-.orderButton {
-  width: 100%;
-  gap: var(--space-8);
-}
-
-.itemCount {
-  background-color: var(--color-surface);
-  color: var(--color-text-gray-900);
-  border-radius: var(--radius-full);
-  width: 20px;
-  height: 20px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 12px;
-  font-weight: bold;
-  flex-shrink: 0;
-}
-
 /* Responsive Design */
-@media (max-width: 1024px) {
-  .container {
-    padding: var(--space-48) var(--space-24);
-  }
-  
-  .content {
-    flex-direction: column;
-  }
-  
-  .sidebar {
-    width: 100%;
-  }
-  
-  .courseNameCell {
-    width: 250px;
-  }
-  
-  .scheduleCell {
-    width: 150px;
-  }
-  
-  .priceCell {
-    width: 100px;
-  }
-}
-
-@media (max-width: 640px) {
-  .container {
-    padding: var(--space-32) var(--space-16);
-  }
-  
-  .row {
-    padding: var(--space-12) var(--space-16);
-  }
-  
-  .courseNameCell {
-    width: 200px;
-  }
-  
-  .pagination {
-    gap: var(--space-4);
-  }
-}
